eslint: Disallow mixed spaces and tabs for indentation
The "no-mixed-spaces-and-tabs" rule seems to be missing from airbnb
linting rules. Consequently, mixed tabulations and spaces in
indentation is not detected by our linter. This was revealed by the
"dev/CLEANUP/IAM_API_OOP" pull request on IronMan-Vault.

This patch adds "no-mixed-spaces-and-tabs" as an error in eslint
configuration. This problem is also currently being reported upstream
to fix airbnb default config:
https://github.com/airbnb/javascript/pull/561.