eslint: Only one newline at end of files
POSIX standard states that the last line of a file should end with a
newline character '\n' [1]. Conventions generally agree on having only
one newline character at the end of files.

This patchs enables a new rule I contributed in eslint [2] to have one
and only one newline at the end of files.

[1]: http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_206
[2]: https://github.com/eslint/eslint/pull/4266